# src/AppForSEII2526.Web/OpenAPIs/swagger2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForSEII2526.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Bocadillo/GetBocadilloParaPedir": {
      "get": {
        "tags": [
          "Bocadillo"
        ],
        "operationId": "GetBocadilloParaPedir",
        "parameters": [
          {
            "name": "filtroTamano",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroTipoPan",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectBocadilloDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectBocadilloDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectBocadilloDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Bocadillo/GetBocadilloParaReseña": {
      "get": {
        "tags": [
          "Bocadillo"
        ],
        "operationId": "GetBocadilloParaReseña",
        "parameters": [
          {
            "name": "filtroNombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroPvp",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectBocadilloDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectBocadilloDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectBocadilloDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SelectBocadilloDTO": {
        "type": "object",
        "properties": {
          "bocadilloID": {
            "type": "integer",
            "format": "int32"
          },
          "nombreBocadillo": {
            "type": "string",
            "nullable": true
          },
          "tamano": {
            "$ref": "#/components/schemas/Tamaño"
          },
          "tipoPanNombre": {
            "type": "string",
            "nullable": true
          },
          "pvp": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
...
      "Tamaño": {
        "enum": [
          "pequeño",
          "normal"
        ],
        "type": "string"
      },
...
    }
  }
}